feat: restructure codebase
