+++ src/const.rs
@@ -1,8 +1,9 @@
 // 128-bit SIMD with u16
-// NOTE: going above 128 bit results in much slower performance
+// NOTE: going above 128 bit results in much slower performance unless AVX2 is enabled at compile
+// time
 pub const SIMD_WIDTH: usize = 16;
 
-// NOTE: be vary carefuly when changing these values since they affect what can fit in
+// NOTE: be very careful when changing these values since they affect what can fit in
 // the u8 scoring without overflowing
 
 pub const MATCH_SCORE: u16 = 6;
@@ -19,5 +20,3 @@ pub const CAPITALIZATION_BONUS: u16 = 4;
 pub const MATCHING_CASE_BONUS: u16 = 2;
 // bonus for haystack == needle
 pub const EXACT_MATCH_BONUS: u16 = 4;
-
-// TODO: bonus for a full continuous match without gaps?
