+++ src/one_shot/bucket.rs
@@ -1,6 +1,8 @@
 use std::marker::PhantomData;
 
-use crate::simd::{smith_waterman, typos_from_score_matrix, SimdNum};
+use crate::smith_waterman::simd::{
+    smith_waterman, typos_from_score_matrix, SimdMask, SimdNum, SimdVec,
+};
 use crate::Match;
 
 pub(crate) trait Bucket<'a> {
@@ -30,8 +32,8 @@ impl<N: SimdNum<L>, const W: usize, const L: usize> FixedWidthBucket<'_, N, W, L
 where
     N: SimdNum<L>,
     std::simd::LaneCount<L>: std::simd::SupportedLaneCount,
-    std::simd::Simd<N, L>: crate::simd::SimdVec<N, L>,
-    std::simd::Mask<N::Mask, L>: crate::simd::SimdMask<N, L>,
+    std::simd::Simd<N, L>: SimdVec<N, L>,
+    std::simd::Mask<N::Mask, L>: SimdMask<N, L>,
 {
     pub fn new() -> Self {
         FixedWidthBucket {
@@ -47,8 +49,8 @@ impl<'a, N: SimdNum<L>, const W: usize, const L: usize> Bucket<'a> for FixedWidt
 where
     N: SimdNum<L>,
     std::simd::LaneCount<L>: std::simd::SupportedLaneCount,
-    std::simd::Simd<N, L>: crate::simd::SimdVec<N, L>,
-    std::simd::Mask<N::Mask, L>: crate::simd::SimdMask<N, L>,
+    std::simd::Simd<N, L>: SimdVec<N, L>,
+    std::simd::Mask<N::Mask, L>: SimdMask<N, L>,
 {
     fn add_haystack(&mut self, haystack: &'a str, idx: usize) {
         assert!(haystack.len() <= W);
