+++ src/smith_waterman/mod.rs
@@ -1,0 +1,2 @@
+pub mod reference;
+pub mod simd;
