+++ src/smith_waterman/simd.rs
